refactor(04-01): extract soft-dep concern to shared/concerns/soft-dep.ts
Move DEPENDENCIES, the Dependency type, the two soft-dep marker constants,
and the soft-dep marker decision out of notify.ts into a new pure
shared/concerns/soft-dep.ts module (MOD-04, D-01 direct-function-call
wiring; no Concern interface, no registry). The central composeReasons
stays in notify.ts with its exact 4-arg signature and delegates its
soft-dep branch via composed.push(...softDepMarkers(...)), so the 13+
orchestrator render-map call sites are untouched.

The concern imports Reason and SoftDepStatus type-only (the cycle
safeguard) and keeps agents-before-mcp marker order. All seven Dependency
importers are repointed to the concern. Output-neutral: catalog-uat stays
4/4 green and docs/output-catalog.md is byte-unchanged.
