+++ extensions/pi-claude-marketplace/orchestrators/plugin/install.ts
@@ -140,9 +140,9 @@ import type { ScopeConfig } from "../../persistence/config-io.ts";
 import type { ScopedLocations } from "../../persistence/locations.ts";
 import type { ExtensionState } from "../../persistence/state-io.ts";
 import type { ExtensionAPI, ExtensionContext } from "../../platform/pi-api.ts";
+import type { Dependency } from "../../shared/concerns/soft-dep.ts";
 import type {
   ContentReason,
-  Dependency,
   PluginFailedMessage,
   PluginInstalledMessage,
   PluginUnavailableMessage,
