+++ extensions/pi-claude-marketplace/orchestrators/plugin/reinstall.ts
@@ -110,9 +110,9 @@ import type { ResolvedPluginInstallable } from "../../domain/resolver.ts";
 import type { ScopedLocations } from "../../persistence/locations.ts";
 import type { ExtensionState } from "../../persistence/state-io.ts";
 import type { ExtensionAPI, ExtensionContext } from "../../platform/pi-api.ts";
+import type { Dependency } from "../../shared/concerns/soft-dep.ts";
 import type {
   ContentReason,
-  Dependency,
   PluginFailedMessage,
   PluginManualRecoveryMessage,
   PluginNotificationMessage,
