+++ extensions/pi-claude-marketplace/orchestrators/reconcile/apply-outcomes.ts
@@ -23,7 +23,8 @@
 // `error.message` in the projection input -- `outcome.reason` is the sole
 // field the renderer reads.
 
-import type { ContentReason, Dependency, Reason } from "../../shared/notify.ts";
+import type { Dependency } from "../../shared/concerns/soft-dep.ts";
+import type { ContentReason, Reason } from "../../shared/notify.ts";
 import type { Scope } from "../../shared/types.ts";
 
 interface OutcomeBase {
